assembly_service/requirements.txt
fastapi
uvicorn
boto3
snowflake-connector-python
snowflake-snowpark-python
python-magic

